report error stack trace to dev & add rate limiter

+++ src/commands/cancel.ts
@@ -1,6 +1,12 @@
 import type { ChatInputCommandInteraction } from "discord.js"
-import { SlashCommandBuilder, EmbedBuilder, ActionRow } from "discord.js"
+import {
+  SlashCommandBuilder,
+  EmbedBuilder,
+  ActionRow,
+  TextChannel,
+} from "discord.js"
 import sql from "../sql"
+import limiter from "../utilities/limiter"
 /**
  * Cancels a pending challenge.
  *
@@ -57,14 +63,18 @@ const cancel = {
       if (interaction.member.roles.cache.has(referee)) {
         const challengeID = interaction.channel.name.split("-")[1]
         const challengeEmbed = challengeLog.isTextBased()
-          ? await challengeLog.messages.fetch(challengeID)
+          ? await limiter.schedule(() =>
+              challengeLog.messages.fetch(challengeID)
+            )
           : undefined
 
         if (!challengeEmbed) {
-          return await interaction.reply({
-            content:
-              "Cannot cancel the challenge since the challenge embed in #challenge-logs does not exist. Contact: <@254643053548142595>",
-          })
+          return await limiter.schedule(() =>
+            interaction.reply({
+              content:
+                "Cannot cancel the challenge since the challenge embed in #challenge-logs does not exist. Contact: <@254643053548142595>",
+            })
+          )
         }
 
         const challengerID = challengeEmbed.embeds?.[0].data.author?.name
@@ -77,10 +87,12 @@ const cancel = {
           interaction.member.id === challengerID ||
           interaction.member.id === opponentID
         ) {
-          await interaction.reply({
-            content: "You cannot cancel a challenge you are already playing.",
-            flags: "Ephemeral",
-          })
+          await limiter.schedule(() =>
+            interaction.reply({
+              content: "You cannot cancel a challenge you are already playing.",
+              flags: "Ephemeral",
+            })
+          )
         } else {
           const challenger = challengerID
             ? interaction.channel.members.get(challengerID)
@@ -98,54 +110,77 @@ const cancel = {
             )
 
           if (!deleteChallengeQuery) {
-            return await interaction.reply({
-              content: `Failed to delete challenge data. Contact: <@254643053548142595>`,
-            })
+            return await limiter.schedule(() =>
+              interaction.reply({
+                content: `Failed to delete challenge data. Contact: <@254643053548142595>`,
+              })
+            )
           }
 
           if (challenger) {
-            await challenger.roles.remove(playing)
+            limiter.schedule(() => challenger.roles.remove(playing))
           }
           if (opponent) {
-            await opponent.roles.remove(playing)
+            await limiter.schedule(() => opponent.roles.remove(playing))
           }
           embed.setFooter({
             text: `Referee cancelled challenge ID: ${challengeID}`,
           })
 
-          await challengeLog.send({ embeds: [embed] })
-          await interaction.reply({
-            content: `Sucessfully cancelled challenge ID: ${challengeID}.`,
-            flags: "Ephemeral",
-          })
-          await interaction.channel.delete()
+          await limiter.schedule(() => challengeLog.send({ embeds: [embed] }))
+          await limiter.schedule(() =>
+            interaction.reply({
+              content: `Sucessfully cancelled challenge ID: ${challengeID}.`,
+              flags: "Ephemeral",
+            })
+          )
+          await limiter
+            .schedule(() => {
+              if (
+                !interaction.channel ||
+                !(interaction.channel instanceof TextChannel)
+              )
+                throw new Error(
+                  "Cancel command used in a non-cached guild or non text-based channel."
+                )
+              return interaction.channel.delete()
+            })
+            .catch(() => null)
         }
       } else {
-        await interaction.reply({
-          content: "You cannot cancel a challenge you are already playing.",
-          flags: "Ephemeral",
-        })
+        await limiter.schedule(() =>
+          interaction.reply({
+            content: "You cannot cancel a challenge you are already playing.",
+            flags: "Ephemeral",
+          })
+        )
       }
     } else {
       // cancelling from any channel
       if (interaction.member.roles.cache.has(playing)) {
-        await interaction.reply({
-          content:
-            "You cannot cancel a challenge when you are already playing.",
-          flags: "Ephemeral",
-        })
+        await limiter.schedule(() =>
+          interaction.reply({
+            content:
+              "You cannot cancel a challenge when you are already playing.",
+            flags: "Ephemeral",
+          })
+        )
       } else if (interaction.member.roles.cache.has(queued)) {
-        await interaction.member.roles.remove(queued)
-        await interaction.reply({
-          content:
-            "You've been removed from the queue of players waiting for a challenge.",
-          flags: "Ephemeral",
-        })
+        await limiter.schedule(() => interaction.member.roles.remove(queued))
+        await limiter.schedule(() =>
+          interaction.reply({
+            content:
+              "You've been removed from the queue of players waiting for a challenge.",
+            flags: "Ephemeral",
+          })
+        )
         embed.setFooter({ text: "Cancelled qeue for challenge" })
-        await challengeLog.send({ embeds: [embed] })
+        await limiter.schedule(() => challengeLog.send({ embeds: [embed] }))
       } else {
         // cancel the first pending challenge
-        const messages = await challengeLog.messages.fetch({ limit: 30 })
+        const messages = await limiter.schedule(() =>
+          challengeLog.messages.fetch({ limit: 30 })
+        )
         const challenges = messages.filter(
           (m) =>
             m.embeds.length === 1 &&
@@ -168,25 +203,36 @@ const cancel = {
               )
 
             if (!deleteChallengeQuery) {
-              return await interaction.reply({
-                content: `Database connection error contact: <@254643053548142595>`,
-              })
+              return await limiter.schedule(() =>
+                interaction.reply({
+                  content: `Database connection error contact: <@254643053548142595>`,
+                })
+              )
             }
 
             embed.setFooter({ text: `Cancelled challenge ID: ${challenge.id}` })
           }
-          await challenge?.delete()
-          await challengeLog.send({ embeds: [embed] })
-          await interaction.reply({
-            content: "Sucessfully cancelled your challenge.",
-            flags: "Ephemeral",
-          })
+          await limiter
+            .schedule(() => {
+              if (!challenge) throw new Error("Challenge channel is undefined")
+              return challenge.delete()
+            })
+            .catch(() => null)
+          await limiter.schedule(() => challengeLog.send({ embeds: [embed] }))
+          await limiter.schedule(() =>
+            interaction.reply({
+              content: "Sucessfully cancelled your challenge.",
+              flags: "Ephemeral",
+            })
+          )
         } else {
-          await interaction.reply({
-            content:
-              "You have not created a challenge and you are not in a queue.",
-            flags: "Ephemeral",
-          })
+          await limiter.schedule(() =>
+            interaction.reply({
+              content:
+                "You have not created a challenge and you are not in a queue.",
+              flags: "Ephemeral",
+            })
+          )
         }
       }
     }
